Add destroy service

+++ src/src/mj_sim.cpp
@@ -126,20 +126,6 @@ static void set_joint_names()
 	}
 }
 
-void load_model()
-{
-	// load and compile model
-	char error[1000] = "Could not load binary model";
-	m = mj_loadXML(tmp_model_path.c_str(), 0, error, 1000);
-	if (!m)
-	{
-		mju_error_s("Could not load model file '%s'", tmp_model_path.c_str());
-	}
-
-	// make data
-	d = mj_makeData(m);
-}
-
 static void init_tmp()
 {
 	// Create directory tmp_mesh_path if not exist
@@ -152,12 +138,15 @@ static void init_tmp()
 	}
 
 	// Copy model meshes to tmp_mesh_path
-	for (const boost::filesystem::directory_entry &file : boost::filesystem::directory_iterator(model_path.parent_path() / model_path_tail))
+	if (boost::filesystem::exists(model_path.parent_path() / model_path_tail))
 	{
-		boost::filesystem::path des_file_path = tmp_mesh_path / file.path().filename();
-		if (!boost::filesystem::exists(des_file_path))
+		for (const boost::filesystem::directory_entry &file : boost::filesystem::directory_iterator(model_path.parent_path() / model_path_tail))
 		{
-			boost::filesystem::copy_file(file.path(), des_file_path);
+			boost::filesystem::path des_file_path = tmp_mesh_path / file.path().filename();
+			if (!boost::filesystem::exists(des_file_path))
+			{
+				boost::filesystem::copy_file(file.path(), des_file_path);
+			}
 		}
 	}
 
@@ -239,22 +228,19 @@ static void add_new_state(mjModel *m_new, mjData *d_new)
 {
 	d_new->time = d->time;
 
-	ROS_INFO("%d", m->nq);
 	int body_id = 0;
 	while (true)
 	{
-		body_id++;
+		body_id++;		
 		const char *name = mj_id2name(m, mjtObj::mjOBJ_BODY, body_id);
 		if (name == nullptr)
 		{
 			break;
-		}		
-		ROS_INFO("%s", name);
+		}
 
-		int body_id_new = mj_name2id(m, mjtObj::mjOBJ_BODY, name);
+		int body_id_new = mj_name2id(m_new, mjtObj::mjOBJ_BODY, name);
 		if (body_id_new == -1)
 		{
-			ROS_WARN("New body [%s] not found", name);
 			continue;
 		}
 
@@ -263,38 +249,42 @@ static void add_new_state(mjModel *m_new, mjData *d_new)
 		{
 			d_new->xfrc_applied[body_id_new + body_nr] = d->xfrc_applied[body_id + body_nr];
 		}
-
+		
 		// Copy joint states
 		int jnt_num = m->body_jntnum[body_id];
 		int jnt_num_new = m_new->body_jntnum[body_id_new];
-		if (jnt_num != jnt_num_new)
+		if (jnt_num == jnt_num_new)
+		{
+			int qpos_adr = m->jnt_qposadr[m->body_jntadr[body_id]];
+			int qpos_adr_new = m_new->jnt_qposadr[m_new->body_jntadr[body_id_new]];
+			for (int jnt_nr = 0; jnt_nr < jnt_num; jnt_nr++)
+			{
+				d_new->qpos[qpos_adr_new + jnt_nr] = d->qpos[qpos_adr + jnt_nr];
+			}
+		}
+		else
 		{
 			ROS_WARN("Old [%s] has %d joints != new [%s] has %d joints", name, jnt_num, name, jnt_num_new);
-			continue;
-		}
-		int jnt_id = m->body_jntadr[body_id];
-		int jnt_id_new = m_new->body_jntadr[body_id_new];
-		for (int jnt_nr = 0; jnt_nr < jnt_num; jnt_nr++)
-		{
-			d_new->qpos[jnt_id_new + jnt_nr] = d->qpos[jnt_id + jnt_nr];
 		}
 
 		// Copy dof states
 		int dof_num = m->body_dofnum[body_id];
 		int dof_num_new = m_new->body_dofnum[body_id_new];
-		if (dof_num != dof_num_new)
+		if (dof_num == dof_num_new)
+		{
+			int dof_adr = m->jnt_dofadr[m->body_jntadr[body_id]];
+			int dof_adr_new = m_new->jnt_dofadr[m_new->body_jntadr[body_id_new]];
+			for (int dof_nr = 0; dof_nr < dof_num; dof_nr++)
+			{
+				d_new->qvel[dof_adr_new + dof_nr] = d->qvel[dof_adr + dof_nr];
+				d_new->qacc_warmstart[dof_adr_new + dof_nr] = d->qacc_warmstart[dof_adr + dof_nr];
+				d_new->qfrc_applied[dof_adr_new + dof_nr] = d->qfrc_applied[dof_adr + dof_nr];
+				d_new->qacc[dof_adr_new + dof_nr] = d->qacc[dof_adr + dof_nr];
+			}
+		}
+		else
 		{
 			ROS_WARN("Old [%s] has %d dofs != new [%s] has %d dofs", name, dof_num, name, dof_num_new);
-			continue;
-		}
-		int dof_id = m->body_dofadr[body_id];
-		int dof_id_new = m_new->body_dofadr[body_id_new];
-		for (int dof_nr = 0; dof_nr < dof_num; dof_nr++)
-		{
-			d_new->qvel[dof_id_new + dof_nr] = d->qvel[dof_id + dof_nr];
-			d_new->qacc_warmstart[dof_id_new + dof_nr] = d->qacc_warmstart[dof_id + dof_nr];
-			d_new->qfrc_applied[dof_id_new + dof_nr] = d->qfrc_applied[dof_id + dof_nr];
-			d_new->qacc[dof_id_new + dof_nr] = d->qacc[dof_id + dof_nr];
 		}
 	}
 
@@ -324,18 +314,115 @@ static void add_new_state(mjModel *m_new, mjData *d_new)
 	m = m_new;
 }
 
-void init_malloc()
+static void init_malloc()
 {
 	MjSim::tau = (mjtNum *)mju_malloc(m->nv * sizeof(mjtNum *));
 	mju_zero(MjSim::tau, m->nv);
+}
+
+static void modify_xml(const char *xml_path, const std::vector<std::string> &remove_body_names = {""})
+{
+	tinyxml2::XMLDocument doc;
+	if (doc.LoadFile(xml_path) != tinyxml2::XML_SUCCESS)
+	{
+		mju_warning_s("Failed to load file \"%s\"\n", xml_path);
+		return;
+	}
+
+	mtx.lock();
+	tinyxml2::XMLElement *worldbody_element = doc.FirstChildElement()->FirstChildElement();
+	std::vector<tinyxml2::XMLNode *> bodies_to_delete;
+	for (tinyxml2::XMLElement *worldbody_element = doc.FirstChildElement()->FirstChildElement();
+			 worldbody_element != nullptr;
+			 worldbody_element = worldbody_element->NextSiblingElement())
+	{
+		if (strcmp(worldbody_element->Value(), "worldbody") == 0)
+		{
+			for (tinyxml2::XMLNode *body_node = worldbody_element->FirstChild();
+					 body_node != nullptr;
+					 body_node = body_node->NextSibling())
+			{
+				if (strcmp(body_node->Value(), "body") == 0)
+				{
+					const char *body_name = body_node->ToElement()->Attribute("name");
+					if (body_name != nullptr && std::find(remove_body_names.begin(), remove_body_names.end(), body_name) != remove_body_names.end())
+					{
+						bodies_to_delete.push_back(body_node);
+					}
+					else if (body_name != nullptr && std::find(MjSim::link_names.begin(), MjSim::link_names.end(), body_name) == MjSim::link_names.end())
+					{
+						int body_id = mj_name2id(m, mjtObj::mjOBJ_BODY, body_name);
+						body_node->ToElement()->SetAttribute("pos",
+																								 (std::to_string(d->xpos[3 * body_id]) + " " +
+																									std::to_string(d->xpos[3 * body_id + 1]) + " " +
+																									std::to_string(d->xpos[3 * body_id + 2]))
+																										 .c_str());
+						body_node->ToElement()->SetAttribute("quat",
+																								 (std::to_string(d->xquat[4 * body_id]) + " " +
+																									std::to_string(d->xquat[4 * body_id + 1]) + " " +
+																									std::to_string(d->xquat[4 * body_id + 2]) + " " +
+																									std::to_string(d->xquat[4 * body_id + 3]))
+																										 .c_str());
+					}
+				}
+			}
+		}
+	}
+	for (tinyxml2::XMLNode *&body_to_delete : bodies_to_delete)
+	{
+		ROS_INFO("Delete body %s", body_to_delete->ToElement()->Attribute("name"));
+		worldbody_element->DeleteChild(body_to_delete);
+	}
+	
+	mtx.unlock();
+
+	doc.SaveFile(xml_path);
+}
+
+bool load_tmp_model(bool reset)
+{
+	char error[1000] = "Could not load binary model";
+	if (reset)
+	{
+		// load and compile model
+		m = mj_loadXML(tmp_model_path.c_str(), 0, error, 1000);
+		if (!m)
+		{
+			mju_warning_s("Could not load model file '%s'", tmp_model_path.c_str());
+			return false;
+		}
+
+		// make data
+		d = mj_makeData(m);
+		init_malloc();
+		return true;
+	}
+	else
+	{
+		// Load current.xml
+		mjModel *m_new = mj_loadXML(tmp_model_path.c_str(), 0, error, 1000);
+		if (!m_new)
+		{
+			mju_warning_s("Load model error: %s", error);
+			return false;
+		}
+
+		mtx.lock();
+		// make data
+		mjData *d_new = mj_makeData(m_new);
+		add_new_state(m_new, d_new);
+		init_malloc();
+		mtx.unlock();
+
+		return true;
+	}
 }
 
 void MjSim::init()
 {
 	set_param();
 	init_tmp();
-	load_model();
-	init_malloc();
+	load_tmp_model(true);
 	set_joint_names();
 	sim_start = d->time;
 }
@@ -346,49 +433,15 @@ bool MjSim::add_data()
 	char error[1000] = "Could not save binary model";
 	mj_saveLastXML(tmp_model_path.c_str(), m, error, 1000);
 
-	// Modify current.xml
+	modify_xml(tmp_model_path.c_str());
+
+	// Add add.xml to current.xml
 	tinyxml2::XMLDocument current_xml_doc;
 	if (current_xml_doc.LoadFile(tmp_model_path.c_str()) != tinyxml2::XML_SUCCESS)
 	{
 		mju_warning_s("Failed to load file \"%s\"\n", tmp_model_path.c_str());
 		return false;
 	}
-	tinyxml2::XMLElement *worldbody_element = current_xml_doc.FirstChildElement()->FirstChildElement();
-	for (tinyxml2::XMLElement *worldbody_element = current_xml_doc.FirstChildElement()->FirstChildElement();
-			 worldbody_element != nullptr;
-			 worldbody_element = worldbody_element->NextSiblingElement())
-	{
-		if (strcmp(worldbody_element->Value(), "worldbody") == 0)
-		{
-
-			for (tinyxml2::XMLElement *body_element = worldbody_element->FirstChildElement();
-					 body_element != nullptr;
-					 body_element = body_element->NextSiblingElement())
-			{
-				if (strcmp(body_element->Value(), "body") == 0)
-				{
-					const char *body_name = body_element->Attribute("name");
-					if (body_name != nullptr && std::find(MjSim::link_names.begin(), MjSim::link_names.end(), body_name) == MjSim::link_names.end())
-					{
-						int body_idx = mj_name2id(m, mjtObj::mjOBJ_BODY, body_name);
-						body_element->SetAttribute("pos",
-																			 (std::to_string(d->xpos[3 * body_idx]) + " " +
-																				std::to_string(d->xpos[3 * body_idx + 1]) + " " +
-																				std::to_string(d->xpos[3 * body_idx + 2]))
-																					 .c_str());
-						body_element->SetAttribute("quat",
-																			 (std::to_string(d->xquat[4 * body_idx]) + " " +
-																				std::to_string(d->xquat[4 * body_idx + 1]) + " " +
-																				std::to_string(d->xquat[4 * body_idx + 2]) + " " +
-																				std::to_string(d->xquat[4 * body_idx + 3]))
-																					 .c_str());
-					}
-				}
-			}
-		}
-	}
-
-	// Add add.xml to current.xml
 	tinyxml2::XMLElement *current_element = current_xml_doc.FirstChildElement();
 	tinyxml2::XMLElement *include_element = current_xml_doc.NewElement("include");
 	include_element->SetAttribute("file", "add.xml");
@@ -398,24 +451,19 @@ bool MjSim::add_data()
 		return false;
 	}
 
-	// Load current.xml
-	mjModel *m_new = mj_loadXML(tmp_model_path.c_str(), 0, error, 1000);
-	if (!m_new)
-	{
-		mju_warning_s("Load model error: %s", error);
-		return false;
-	}
+	return load_tmp_model(false);
+}
 
-	mtx.lock();
-	// make data
-	mjData *d_new = mj_makeData(m_new);
+bool MjSim::remove_body(const std::vector<std::string> &body_names)
+{
+	// Save current.xml
+	char error[1000] = "Could not save binary model";
+	mj_saveLastXML(tmp_model_path.c_str(), m, error, 1000);
 
-	add_new_state(m_new, d_new);
+	// Modify current.xml
+	modify_xml(tmp_model_path.c_str(), body_names);
 
-	init_malloc();
-	mtx.unlock();
-
-	return true;
+	return load_tmp_model(false);
 }
 
 void MjSim::controller()
